Добавил сервис для управления содержимым страницы "Контакты"

+++ OxyNode/Services/ContactsService.cs
@@ -54,8 +54,45 @@ namespace OxyNode.Services
         // Read
         public async Task<Contacts> ReadContacts()
         {
-            // поскольку документ 1, ожно получить пустым Find
-            return await ContactsCollection.Find(new BsonDocument()).FirstOrDefaultAsync();
+            if (HasContacts == true)
+            {
+                // поскольку документ 1, ожно получить пустым Find
+                return await ContactsCollection.Find(new BsonDocument()).FirstOrDefaultAsync();
+            }
+            // если "Контакты" не созданы - вернуть заглушку
+            else
+            {
+                return new Contacts()
+                {
+                    CompanyName = "_Имя_компании",
+                    PhoneNumbers = new List<string>()
+                    {
+                        "8(123)4567890",
+                        "8(234)5678910"
+                    },
+                    Email = "info@имя_компании.com",
+
+                    AdditionalInfoKeys = new List<string>()
+                    {
+                        "ОГРН: ",
+                        "Юридический адрес: ",
+                        "Почтовый адрес: ",
+                        "Факсы: ",
+                        "Время работы: "
+                    },
+                    AdditionalInfoValues = new List<string>()
+                    {
+                        "123456789",
+                        "Планета Земля, Российская Федерация, Подольск",
+                        "Планета Земля, Российская Федерация, Домодедово",
+                        "поют романсы...",
+                        "24/7/365"
+                    },
+                    PathToScheme = "<div style=\"position:relative; overflow: hidden; \"><a href=\"https://yandex.ru/maps/org/mospolitekh/1701016484/?utm_medium=mapframe&utm_source=maps\" style=\"color:#eee;font-size:12px;position:absolute;top:0px;\">Мосполитех</a><a href=\"https://yandex.ru/maps/213/moscow/category/university_college/?utm_medium=mapframe&utm_source=maps\" style=\"color:#eee;font-size:12px;position:absolute;top:14px;\">ВУЗ в Москве</a><iframe src=\"https://yandex.ru/map-widget/v1/-/CKce5ROb\" width=\"560\" height=\"400\" frameborder=\"1\" allowfullscreen=\"true\" style=\"position:relative;\"></iframe></div>"
+                };
+            }
+
+            
         }
 
         // Update
@@ -77,8 +114,12 @@ namespace OxyNode.Services
         // Delete
         public async Task DeleteContacts()
         {
-            HasContacts = false;    // установка флага, что "Контактов" нет
-            await ContactsCollection.DeleteOneAsync(new BsonDocument());
+            // Если "Контакты" есть - удалить их
+            if (HasContacts == true)
+            {
+                HasContacts = false;    // установка флага, что "Контактов" нет
+                await ContactsCollection.DeleteOneAsync(new BsonDocument());
+            }
         }
 
         #endregion
